reorg things, move more CF stuff into lib.rs
